separate client list into a component
